Small math fix + revert to float, which wasn't causing problems after all

+++ libhfst/src/parsers/pmatch_utils.h
@@ -35,7 +35,7 @@ struct PmatchObject;
 struct PmatchTransducerContainer;
 
 typedef std::pair<std::string, std::string> StringPair;
-typedef double WordVecFloat;
+typedef float WordVecFloat;
 struct WordVector;
 
 extern char* data;
